# src/main/resources/openapi/jobly_workflow_openapi.yaml
openapi: 3.0.4
info:
  title: Jobly API
  description: |-
    This service will be used to manage the Jobly application.
  version: 1.0.0
servers:
  - url: https://jobly.com/api/v1
paths:
...
  /applications/mine:
    get:
      summary: Get applications submitted by the authenticated user
      tags:
        - Application
      operationId: getMyApplications
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: A list of applications submitted by the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMyApplicationsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /applications/mine/{id}:
    get:
      summary: Get application details by ID for the authenticated user
      tags:
        - Application
      operationId: getMyApplication
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the application to retrieve
      responses:
        '200':
          description: A list of applications submitted by the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyApplication'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /applications/{id}/cancel:
    post:
      summary: Cancel an application by ID for the authenticated user
      tags:
        - Application
      operationId: cancelApplication
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the application to cancel
      responses:
        '204':
          description: Application cancelled successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /applications/{id}:
    put:
      summary: Update an application by ID for the authenticated user
      tags:
        - Application
      operationId: updateApplication
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the application to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationUpdateRequest'
      responses:
        '200':
          description: Application updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
  /applications/{id}/manage:
    post:
      summary: Manage an application (e.g., approve or reject) for job offer owner
      tags:
        - Application
      operationId: manageApplication
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the application to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationManageRequest'
      responses:
        '204':
          description: Application managed successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedMessage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorMessage'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    DateTime:
      type: string
      format: date-time
      example: '2023-10-01T12:00:00Z'
    SystemGeneratedId:
      type: integer
      format: int64
      example: 1
    BadRequestMessage:
      type: object
      properties:
        status:
          type: integer
          example: 400
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: Invalid input data
    UnauthorizedMessage:
      type: object
      properties:
        status:
          type: integer
          example: 401
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: User is unauthorized to access this resource
    ForbiddenMessage:
      type: object
      properties:
        status:
          type: integer
          example: 403
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: User is forbidden from accessing this resource
    NotFoundMessage:
      type: object
      properties:
        status:
          type: integer
          example: 404
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: Resource not found
...
    InternalServerErrorMessage:
      type: object
      properties:
        status:
          type: integer
          example: 500
        timestamp:
          $ref: '#/components/schemas/DateTime'
        message:
          type: string
          example: An unexpected error occurred on the server
...
    JobOfferListObject:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        title:
          type: string
          example: Senior Java Developer
          maxLength: 32
        companyName:
          type: string
          example: Tech Solutions Inc.
          maxLength: 60
        salary:
          type: number
          format: float
          example: 85000.00
        workType:
          $ref: '#/components/schemas/WorkType'
        location:
          type: string
          example: New York, NY
          maxLength: 100
        category:
          $ref: '#/components/schemas/JobOfferCategory'
...
    WorkType:
      type: string
      enum: [ ON_SITE, REMOTE, HYBRID ]
...
    JobOfferCategory:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        name:
          type: string
          example: Software Development
          maxLength: 32
...
    Application:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        comment:
          type: string
          maxLength: 1000
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        applicant:
          $ref: '#/components/schemas/Applicant'
        applicationStatus:
          $ref: '#/components/schemas/ApplicationStatus'
        cvId:
          $ref: '#/components/schemas/SystemGeneratedId'
    ApplicationStatus:
      type: string
      enum: [ PENDING, ACCEPTED, REJECTED, WITHDRAWN ]
    Applicant:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        firstName:
          type: string
          example: Jane
          maxLength: 20
        lastName:
          type: string
          example: Smith
          maxLength: 20
        email:
          type: string
          format: email
...
    GetMyApplicationsResponse:
      type: object
      properties:
        total:
          type: integer
          example: 3
        applications:
          type: array
          items:
            $ref: '#/components/schemas/MyApplicationListObject'
    MyApplicationListObject:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        jobOffer:
          $ref: '#/components/schemas/JobOfferListObject'
        applicationStatus:
          $ref: '#/components/schemas/ApplicationStatus'
    MyApplication:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SystemGeneratedId'
        comment:
          type: string
          maxLength: 1000
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        jobOffer:
          $ref: '#/components/schemas/JobOfferListObject'
        cvId:
          $ref: '#/components/schemas/SystemGeneratedId'
    ApplicationUpdateRequest:
      type: object
      properties:
        comment:
          type: string
          example: I have updated my application with more relevant experience.
          maxLength: 1000
    ApplicationManageRequest:
      type: object
      properties:
        action:
          $ref: '#/components/schemas/ApplicationManageAction'
    ApplicationManageAction:
      type: string
      enum: [ APPROVE, REJECT ]